Fixed Multi Echo
FINALLY! Needed to remove 1 + sign... That was essentially it. I hate
everything... Except for these echos. Cause they are awesome.
