+++ Assets/Scripts/EchoSphere.cs
@@ -1,86 +1,57 @@
 using UnityEngine;  
 using System;  
 using System.Collections;
-using System.Collections.Generic;
-
 
 public class EchoSphere : MonoBehaviour {  
 	// Echo sphere Properties
-	public List<int> Spheres = new List<int>();
-
 	public Material EchoMaterial = null;
 	public Texture EchoTexture = null;
+	//public Shader EchoShader = null;
 	
-	public float SphereMaxRadius = 40.0f;     //Final size of the echo sphere.
-	public float SphereCurrentRadius = 0.0f;  //Current size of the echo sphere
+	public float SphereMaxRadius = 10.0f;     //Final size of the echo sphere.
+	public float CurrentRadius = 0.0f;  //Current size of the echo sphere
 	
 	public float FadeDelay = 0.0f;          //Time to delay before triggering fade.
 	public float FadeRate = 1.0f;           //Speed of the fade away
-	public float EchoSpeed = 9.0f;          //Speed of the sphere growth.
-
+	public float EchoSpeed = 1.0f;          //Speed of the sphere growth.
+	
 	public int SphereCount = 1;
 	public int CurrentSphere = 0;
-	public int SphereIndex = 0;
-
+	
 	private bool isAnimated = false;    
 	private float deltaTime = 0.0f;
 	
 	public float fade = 0.0f;
 	public bool isTexturedScene = true;
-
-	public Vector3 Position;
-
 	public Vector3 pingLocation;
 	public Vector3 rockLocation;
 	public TestingNightVision appControl;
 	public bool isGrounded;
 	public RockNoise rockNoise;
 	public GameObject rock;
-	//public EchoSphere2 es;
-
+	
 	// Use this for initialization
 	void Start () 
 	{	
-		//SetupSimpleScene1();
-		InitializeSpheres();
-		//es = GameObject.FindGameObjectWithTag ("Player").GetComponent<EchoSphere2> ();
+		SetupSimpleScene1();
 		appControl = GetComponent<TestingNightVision> ();
 	}
 	
 	/// 
 	/// Scenario1: Monocolor echo. 
-	///
-	void InitializeSpheres()
-	{
-		for (int i = 0; i < SphereCount; i++) 
-		{
-			//EchoSphere2 es = new  EchoSphere2
-			{
-				EchoMaterial = EchoMaterial;
-				EchoTexture = EchoTexture;
-				EchoSpeed = EchoSpeed;
-				SphereMaxRadius = SphereMaxRadius;
-				FadeDelay = FadeDelay;
-				FadeRate = FadeRate;
-				SphereIndex = i;
-				isTexturedScene = false;
-			};
-			Spheres.Add(1);
-		}
-	}
-
+	/// 
 	void SetupSimpleScene1()
 	{
 		SphereMaxRadius = 40.0f;
-		SphereCurrentRadius = 0.0f;
+		CurrentRadius = 0.0f;
 		FadeDelay = 0.0f;
 		FadeRate = 1f;
 		EchoSpeed = 9.0f;
 		EchoMaterial.mainTexture = null;
 		//EchoShader = Shader.Find ("GlowOutline");
 		//rend = GetComponent<Renderer> ();
-
-		//EchoMaterial.SetFloat("_DistanceFade",1.0f);
+		
+		EchoMaterial.SetFloat("_DistanceFade",1.0f);
 		isTexturedScene = false;
 	}
 	
@@ -90,17 +61,17 @@ public class EchoSphere : MonoBehaviour {
 	void SetupSimpleScene2()
 	{
 		SphereMaxRadius = 40.0f;
-		SphereCurrentRadius = 0.0f;
+		CurrentRadius = 0.0f;
 		FadeDelay = 0.0f;
 		FadeRate = 1.0f;
 		EchoSpeed = 9.0f;
 		EchoMaterial.mainTexture = EchoTexture;
-
-		//EchoMaterial.SetFloat("_DistanceFade",0.0f);
+		
+		EchoMaterial.SetFloat("_DistanceFade",0.0f);
 		isTexturedScene = true;
 	}
 	
-/*	void OnGUI () {
+	/*	void OnGUI () {
 		// Make a background box
 		GUI.Box(new Rect(10,10,100,90), "Scenarios");
 		
@@ -133,23 +104,16 @@ public class EchoSphere : MonoBehaviour {
 			//rock = null;
 		}
 		deltaTime += Time.deltaTime;
-
-		if(EchoMaterial == null)return;	
-		foreach (int es in Spheres)
-		{
-			//es.Update();
-		}
+		
 		UpdateRayCast();
 		UpdateEcho();
-		UpdateProperties ();
-		//UpdateShader();
+		UpdateShader();
 	}
 	
 	// Called to trigger an echo pulse
-	void TriggerPulse()
-	{
+	void TriggerPulse(){
 		deltaTime = 0.0f;
-		SphereCurrentRadius = 0.0f;
+		CurrentRadius = 0.0f;
 		fade = 0.0f;
 		isAnimated = true;
 		if(rockNoise.isgrounded)
@@ -167,7 +131,7 @@ public class EchoSphere : MonoBehaviour {
 	void ClearPulse()
 	{
 		fade = 0.0f;
-		SphereCurrentRadius = 0.0f;
+		CurrentRadius = 0.0f;
 		isAnimated = false;
 	}
 	
@@ -180,14 +144,10 @@ public class EchoSphere : MonoBehaviour {
 			RaycastHit hit;
 			if (Physics.Raycast(ray,out hit, Mathf.Infinity)) 
 			{
-				Debug.Log("Triggering pulse["+CurrentSphere.ToString()+"]");
 				pingLocation = gameObject.transform.position;
-				//Spheres[CurrentSphere] = hit.point;
 				EchoMaterial.SetVector("_Position", hit.point);
-				//Spheres[CurrentSphere].TriggerPulse();
+				//rend.material.shader = EchoShader;
 				TriggerPulse();
-				CurrentSphere += 1;
-				if(CurrentSphere >= Spheres.Count)CurrentSphere = 0;
 			}
 		}  
 		if (Input.GetButtonDown("Fire2") && appControl.isSonar){
@@ -197,9 +157,8 @@ public class EchoSphere : MonoBehaviour {
 			{
 				pingLocation = gameObject.transform.position;
 				EchoMaterial.SetVector("_Position", pingLocation);
-			//	Spheres[CurrentSphere].Position = hit.point;
+				//rend.material.shader = EchoShader;
 				TriggerPulse();
-				//Spheres[CurrentSphere].TriggerPulse();
 			}
 		}  
 		if(rockNoise.isgrounded)
@@ -212,54 +171,19 @@ public class EchoSphere : MonoBehaviour {
 	}
 	
 	// Called to update the echo front edge
-	void UpdateEcho()
-	{
+	void UpdateEcho(){
 		if(!isAnimated)return;
-		if(SphereCurrentRadius >= SphereMaxRadius)
-		{
-			HaltPulse();
-		} else 
-		{
-			SphereCurrentRadius += Time.deltaTime * EchoSpeed;  
-		}
 		
-		float radius = SphereCurrentRadius;
-		float maxRadius = SphereMaxRadius;
-		float maxFade = SphereMaxRadius / EchoSpeed;
-		if(fade > maxFade)
-		{
-			return;
-		}
-		
-		if(deltaTime > FadeDelay)
-			fade += Time.deltaTime * FadeRate;
-		/*
 		if(CurrentRadius >= SphereMaxRadius){
 			HaltPulse();
 		} else {
 			CurrentRadius += Time.deltaTime * EchoSpeed;  
-		}*/
+		}
 	}
-
-	void UpdateProperties()
-	{
-		if(!isAnimated)return;
-		float maxRadius = SphereMaxRadius;
-		float maxFade = SphereMaxRadius / EchoSpeed;
-		
-		Debug.Log("Updating _Position"+SphereIndex.ToString());
-		EchoMaterial.SetVector("_Position"+SphereIndex.ToString(),Position);
-		EchoMaterial.SetFloat("_Radius"+SphereIndex.ToString(),SphereCurrentRadius);
-		EchoMaterial.SetFloat("_Fade"+SphereIndex.ToString(),fade);
-		
-		EchoMaterial.SetFloat("_MaxRadius",maxRadius);
-		EchoMaterial.SetFloat("_MaxFade",maxFade);
-	}
-
+	
 	// Called to update the actual shader values (some of which only change once but are included here
 	// for illustrative purposes)
-	/*void UpdateShader()
-	{
+	void UpdateShader(){
 		float radius = CurrentRadius;
 		float maxRadius = SphereMaxRadius;
 		float maxFade = SphereMaxRadius / EchoSpeed;
@@ -273,5 +197,5 @@ public class EchoSphere : MonoBehaviour {
 		EchoMaterial.SetFloat("_MaxRadius",maxRadius);
 		EchoMaterial.SetFloat("_Fade",fade);
 		EchoMaterial.SetFloat("_MaxFade",maxFade);
-	}*/
+	}
 }
